chore: Set up initial src folder structure, moved APP.tsx and insalled navigation libraries
