+++ __tests__/App.test.tsx
@@ -4,7 +4,7 @@
 
 import React from 'react';
 import ReactTestRenderer from 'react-test-renderer';
-import App from '../App';
+import App from '../src/App';
 
 test('renders correctly', async () => {
   await ReactTestRenderer.act(() => {
